+++ game-crimsondesert/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - None Planned
 
+## [0.2.2] - 2026-03-28
+
+- Changed: Patch mod installer updated for new modding patterns (JSON/QT Mod Managers) - "0036+" folders
+- Added: Separate installer for JSON mods with no data files
+- Added: Support for QT Mod Manager
+- Removed: CD Mod Manager tool and download button (not currently available)
+
 ## [0.2.1] - 2026-03-26
 
 - Added: Download buttons and setup scripts for tools